feat: add poseidon encryption implementation
