+++ manta-pay/src/crypto/mod.rs
@@ -19,5 +19,5 @@
 pub mod constraint;
 pub mod ecc;
 pub mod encryption;
-pub mod hash;
 pub mod key;
+pub mod poseidon;
